pyproject.toml
[tool.poetry]
name = "renewable-energy-consultant"
version = "0.1.0"
description = "An AI-powered renewable energy consultant with dashboard capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "mcp_server" },
    { include = "api" }
]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
flask-socketio = "^5.3.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pandas = "^2.1.1"
numpy = "^1.25.2"
matplotlib = "^3.8.0"
seaborn = "^0.12.2"
scikit-learn = "^1.3.0"
openai = "^1.3.0"
anthropic = "^0.8.1"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
sqlalchemy = "^2.0.21"
plotly = "^5.18.0"
beautifulsoup4 = "^4.12.2"
aiohttp = "^3.8.5"
eventlet = "^0.33.3"
pymongo = "^4.5.0"
hypercorn = "^0.14.4"
mcp-client-python = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 